+++ src/Home.css
@@ -1,31 +1,44 @@
-.home {
+/* .home {
     background-color: white;
     position: relative;
     font-family: 'Roboto', sans-serif;
     font-style: normal;
     display: flex;
-    flex-direction: column;
-}
+    flex-direction: row;
+    width: 100%;
+    height: 100%;
+} */
 
 .home-wrapper {
     display: flex;
-    align-items: baseline;
-    justify-content: baseline;
-    flex-direction: row;
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 32px;
+    margin-left: 400px;
 }
 
-.home-list {
+.home {
+    position: relative;
+    font-family: 'Roboto', sans-serif;
+    font-style: normal;
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    height: auto;
+    align-items: flex-start;
+}
+
+/* .home-list {
     display: flex;
     align-items: baseline;
     justify-content: baseline;
     flex-direction: column;
-}
+} */
 
 .par-1 {
     font-weight: 800;
     font-size: 32px;
     line-height: 36px;
-    margin: 10px;
 }
 
 .par-2 {
