+++ src/components/ComponentsOverview.css
@@ -1,4 +1,4 @@
-.checkbox-wrapper {
+.components-wrapper {
     display: flex;
     flex-wrap: wrap;
     align-items: center;
