+++ src/components/SliderbarMenu.css
@@ -1,27 +1,32 @@
 .sidebar {
     display: flex;
-    align-items: center;
-    justify-content: baseline;
     flex-direction: column;
     background-color: #f0f0f0;
     width: 300px;
-    height: 1000px;
+    height: 100vh;
+    position: fixed;
+    left: 0;
+    top: 0;
+    padding: 20px;
   }
   
   .sidebar ul {
     list-style-type: none;
+    padding: 0;
+    margin: 0;
   }
   
   .sidebar li {
-    margin-bottom: 32px;
-    width: 200px;
-    height: 32px;
-    font-weight: 400;
+    margin-bottom: 16px;
   }
   
-  .sidebar li:hover {
-    background-color: #c4bfbf;
-    width: 100%;
-    height: 32px;
+  .sidebar li a {
+    text-decoration: none;
+    color: black;
+    font-weight: bold;
+    font-size: 16px;
   }
   
+  .sidebar li a:hover {
+    color: #c4bfbf;
+  }
