+++ src/components/cards/CardWrapper.css
@@ -1,9 +1,13 @@
 .card {
     display: flex;
     flex-direction: row;
-    width: 100%;
     gap: 32px;
-    justify-content: center;
+    justify-content: flex-start;
     align-items: center;
-    flex-wrap: wrap;
+}
+
+@media (max-width: 50%) {
+    .card {
+        flex-direction: column;
+    }
 }
